TST: collect together the tests for support edges & out-of-bounds
