+++ scipy/stats/tests/test_discrete_basic.py
@@ -9,7 +9,7 @@ from scipy.lib.six import xrange
 from scipy import stats
 from common_tests import (check_normalization, check_moment, check_mean_expect,
         check_var_expect, check_skew_expect, check_kurt_expect,
-        check_entropy, check_private_entropy)
+        check_entropy, check_private_entropy, check_edge_support)
 
 DECIMAL_meanvar = 0  # 1  # was 0
 
@@ -59,6 +59,8 @@ def test_discrete_basic():
         yield check_discrete_chisquare, distfn, arg, rvs, alpha, \
                       distname + ' chisquare'
 
+        yield check_edge_support, distfn, arg
+
     seen = set()
     for distname, arg in distdiscrete:
         if distname in seen:
@@ -96,16 +98,6 @@ def test_moments():
         cond = distname in ['zipf']
         msg = distname + ' fails kurtosis'
         yield knf(cond, msg)(check_kurt_expect), distfn, arg, m, v, k, distname
-
-
-@npt.dec.slow
-def test_discrete_extra():
-    for distname, arg in distdiscrete:
-        distfn = getattr(stats,distname)
-        yield check_ppf_limits, distfn, arg, distname + \
-              ' ppf limit test'
-        yield check_isf_limits, distfn, arg, distname + \
-              ' isf limit test'
 
 
 @npt.dec.skipif(True)
@@ -202,39 +194,6 @@ def check_oth(distfn, arg, msg):
     npt.assert_(distfn.sf(median_sf - 1, *arg) > 0.5)
     npt.assert_(distfn.cdf(median_sf + 1, *arg) > 0.5)
     npt.assert_equal(distfn.isf(0.5, *arg), distfn.ppf(0.5, *arg))
-
-# next 3 functions copied from test_continous_extra
-#    adjusted
-
-
-def check_ppf_limits(distfn,arg,msg):
-    below,low,upp,above = distfn.ppf([-1,0,1,2], *arg)
-    # print distfn.name, distfn.a, low, distfn.b, upp
-    # print distfn.name,below,low,upp,above
-    assert_equal_inf_nan(distfn.a-1,low, msg + 'ppf lower bound')
-    assert_equal_inf_nan(distfn.b,upp, msg + 'ppf upper bound')
-    npt.assert_(np.isnan(below), msg + 'ppf out of bounds - below')
-    npt.assert_(np.isnan(above), msg + 'ppf out of bounds - above')
-
-
-def check_isf_limits(distfn,arg,msg):
-    below,low,upp,above = distfn.isf([-1,0,1,2], *arg)
-    # print distfn.name, distfn.a, low, distfn.b, upp
-    # print distfn.name,below,low,upp,above
-    assert_equal_inf_nan(distfn.a-1,upp, msg + 'isf lower bound')
-    assert_equal_inf_nan(distfn.b,low, msg + 'isf upper bound')
-    npt.assert_(np.isnan(below), msg + 'isf out of bounds - below')
-    npt.assert_(np.isnan(above), msg + 'isf out of bounds - above')
-
-
-def assert_equal_inf_nan(v1,v2,msg):
-    npt.assert_(not np.isnan(v1))
-    if not np.isinf(v1):
-        npt.assert_almost_equal(v1, v2, decimal=10, err_msg=msg +
-                                   ' - finite')
-    else:
-        npt.assert_(np.isinf(v2) or np.isnan(v2),
-               msg + ' - infinite, v2=%s' % str(v2))
 
 
 def check_sample_skew_kurt(distfn, arg, sk, ss, msg):
